Fix: free/total disk size on x32 systems (#332)

+++ daemon/util/FileSystem.h
@@ -31,8 +31,8 @@ class FileSystem
 {
 	struct DiskState
 	{
-		size_t available;
-		size_t total;
+		int64 available;
+		int64 total;
 	};
 
 public:
